Use one single component for drop areas
